This is a (rather) large, messy commit. Goals:
* It moves the logic for Memory and RegStorage out of the context. A
context is now generic over 2 types: memory(which implements trait Memory)
and register storage(which implements RegStore).

* Memory: Since we define a trait for memory to be implemented, this
allows us to add different models for symbolic memory. In seg_mem.rs, I
have started with the basic idea for implementing the segment-tree
backed idea is going to be worked upon soon. In qword_mem.rs, the
initial memory model of chunks of size of a qword(8 bytes) is present.

* RegStore: This allows us to decouple register store/load logic from
the context core.

* The directory structure is kept such that we have ->
    - module/
        - trait_definition
        - [trait implementation - 1]
        - [trait implementation - 2]
        ..
    - utils/